Major update add admin authentication, 2 factor authentication  and fix some bugs

+++ Backened/controllers/foodController.js
@@ -1,5 +1,6 @@
 import foodModel from '../models/foodModel.js'
 import fs from 'fs'
+import userModel from '../models/userModel.js';
 
 //add food item
 
@@ -13,8 +14,12 @@ const addFood = async (req,res) => {
         image:image_filename
     })
     try{
+        let userData = await userModel.findById(req.body.userId);
+        if(userData && userData.role=== "admin"){
         await food.save();
-        res.json({success:true,message:"Food added"})
+        }else{
+            res.json({success:false,message:"You are not admin"})
+        }
     }catch(error){
         console.log(error)
         res.json({success:false,message:"Error"})
@@ -36,12 +41,16 @@ const listFood = async(req,res) =>{
 //remove food item
 const removeFood = async(req,res) =>{
     try{
-        const food =await foodModel.findById(req.body.id);
-        fs.unlink(`uploads/${food.image}`,()=>{})
+        const userData = await userModel.findById(req.body.userId);
+        if(userData && userData.role === "admin"){
+            const food =await foodModel.findById(req.body.id);
+            fs.unlink(`uploads/${food.image}`,()=>{})
 
-        await foodModel.findByIdAndDelete(req.body.id);
-        res.json({success:true,message:"Food removed"})
-
+            await foodModel.findByIdAndDelete(req.body.id);
+            res.json({success:true,message:"Food removed"})
+        }else{
+            res.json({success:false,message:"You are not admin"})
+        } 
     }catch(error){
         console.log(error);
         res.json({success:false,message:"error"})
